src/main/java/io/pivotal/pal/tracker/TimeEntryHealthIndicator.java
package io.pivotal.pal.tracker;

import org.springframework.boot.actuate.health.Health;
import org.springframework.boot.actuate.health.HealthIndicator;
import org.springframework.stereotype.Component;

@Component
public class TimeEntryHealthIndicator implements HealthIndicator {

    private TimeEntryRepository timeEntryRepository;

    public TimeEntryHealthIndicator(TimeEntryRepository timeEntryRepository) {
        this.timeEntryRepository = timeEntryRepository;
    }

    @Override
    public Health health() {
        if (isHealthy()) {
            return Health.up().build();
        }

        return Health.down().build();
    }

    private boolean isHealthy() {
        return this.timeEntryRepository.list().size() < 5;
    }

}
